+++ cargo-rbmt/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Add version check at runtime if `workspace.metadata.rbmt.version` is set to a semver or hash value to ensure the right version of `cargo-rbmt` is running.
+
 # [0.4.1] - 2026-06-24
 
 * Fix windows support with correct internal `Path` usage.